ml/requirements.txt
numpy>=1.19.2
opencv-python>=4.5.1
Pillow>=8.2.0
tensorflow>=2.5.0
tqdm>=4.60.0
scipy>=1.7.0
matplotlib>=3.3.0 